+++ src/private/release/InvokeNovaPublishWorkflow.ps1
@@ -1,3 +1,26 @@
+function Test-NovaPublishWorkflowShouldImportLocalModule {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][pscustomobject]$WorkflowContext,
+        [switch]$ShouldRun
+    )
+
+    return $ShouldRun -and $null -ne $WorkflowContext.LocalPublishActivation
+}
+
+function Invoke-NovaPublishWorkflowCiRestore {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][pscustomobject]$WorkflowContext,
+        [switch]$ShouldRun,
+        [switch]$ContinuousIntegrationRequested
+    )
+
+    if ($ShouldRun -and $ContinuousIntegrationRequested) {
+        $null = Import-NovaBuiltModuleForCi -ProjectInfo $WorkflowContext.ProjectInfo
+    }
+}
+
 function Invoke-NovaPublishWorkflow {
     [CmdletBinding()]
     param(
@@ -6,15 +29,20 @@ function Invoke-NovaPublishWorkflow {
     )
 
     $publishParams = $WorkflowContext.PublishParams
+    $continuousIntegrationRequested = ($WorkflowContext.PSObject.Properties.Name -contains 'ContinuousIntegrationRequested') -and $WorkflowContext.ContinuousIntegrationRequested
+    $shouldImportPublishedLocalModule = Test-NovaPublishWorkflowShouldImportLocalModule -WorkflowContext $WorkflowContext -ShouldRun:$ShouldRun
 
     Invoke-NovaBuildValidation -WorkflowContext $WorkflowContext
 
     & $WorkflowContext.PublishInvocation.Action @publishParams
 
-    if (-not $ShouldRun -or -not $WorkflowContext.LocalPublishActivation) {
+    if (-not $shouldImportPublishedLocalModule) {
+        Invoke-NovaPublishWorkflowCiRestore -WorkflowContext $WorkflowContext -ShouldRun:$ShouldRun -ContinuousIntegrationRequested:$continuousIntegrationRequested
         return
     }
 
     $null = & $WorkflowContext.LocalPublishActivation.ImportAction -ProjectName $WorkflowContext.PublishInvocation.Parameters.ProjectInfo.ProjectName -ManifestPath $WorkflowContext.LocalPublishActivation.ManifestPath
     Write-Verbose "Module copy to local path complete and imported from $( $WorkflowContext.LocalPublishActivation.ManifestPath )"
+
+    Invoke-NovaPublishWorkflowCiRestore -WorkflowContext $WorkflowContext -ShouldRun:$ShouldRun -ContinuousIntegrationRequested:$continuousIntegrationRequested
 }